09/16/2014 usrivastava [BUG] #1296 Fixed unit of measure null error
git-svn-id: https://www.etbmanager.org/svn/pharmadex/pharmadex/trunk@165 24bb7120-ba24-2248-a177-eda5b6c43647

+++ core/src/main/java/org/msh/pharmadex/dao/iface/DosUomDAO.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Author: usrivastava
  */
-public interface DosUomDAO extends JpaRepository<DosUom, Long> {
+public interface DosUomDAO extends JpaRepository<DosUom, Integer> {
 
 
 }
